Final version, features tunable parameters
The robot is under a mask that allows for the simple tuning of parameters and outputs the resultant joint torques. Since the robot is moving in a vacuum, the joint torques may not be accurate. However, in a realistic physical environment, this motion would result in actual motion due to frictional forces.